lint

+++ include/autd3/driver/firmware_version.hpp
@@ -3,7 +3,7 @@
 // Created Date: 10/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 15/11/2022
+// Last Modified: 26/11/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -73,10 +73,10 @@ struct FirmwareInfo {
   [[nodiscard]] static std::string firmware_version_map(const uint8_t version_num) {
     if (version_num == 0) return "older than v0.4";
     if (version_num <= 0x06) return "v0." + std::to_string(version_num + 3);
-    if (version_num <= 0x09) return "unknown (" + std::to_string(static_cast<int>(version_num)) + ")";
+    if (version_num <= 0x09) return "unknown (" + std::to_string(version_num) + ")";
     if (version_num <= 0x15) return "v1." + std::to_string(version_num - 0x0A);
     if (version_num <= 0x86) return "v2." + std::to_string(version_num - 0x80);
-    return "unknown (" + std::to_string(static_cast<int>(version_num)) + ")";
+    return "unknown (" + std::to_string(version_num) + ")";
   }
 
  private:
